+++ Kikin-Informed-Search-Algorithms/WeightedAStar.py
@@ -15,10 +15,10 @@ def WeightedAStarSearch(graph, heuristics, origin, destination):
             break
 
         for next_node in graph.get_neighbors(current):
-            new_cost = cost_so_far[current] + weight * (int(graph.get_weight(current, next_node)))
+            new_cost = cost_so_far[current] + weight * (float(graph.get_weight(current, next_node)))
             if next_node not in cost_so_far or new_cost < cost_so_far[next_node]:
                 cost_so_far[next_node] = new_cost
-                priority = new_cost + weight * (int(heuristics.get_weight(next_node, destination)))
+                priority = new_cost + weight * (float(heuristics.get_weight(next_node, destination)))
                 frontier.put(next_node, priority)
                 came_from[next_node] = current
